🚚 rename event modules into events
